Circular array ,only head pointer ,no tail pointer
The problem is that when a circular array Q [0.. maxlen-1] is used to represent a queue, the queue has only one head pointer front and no tail pointer. Instead, a counter count is set to record the number of nodes in the queue. Try to write an algorithm that can initialize the queue, judge the queue empty, read the team header elements, join the team operation and exit the team operation.

When entering the team, judge whether the team is full or not, and output the prompt message queue is full when the team is full!

When leaving the team and reading the team head element, judge whether the team is empty, and output the prompt message queue is empty!